aivi-artifisv2-robot/requirements.txt
pika==1.2.0
azure-storage-blob
pandas==1.1.1
pyftpdlib==1.5.6
prometheus-client~=0.11.0
